set up proper inference code

+++ convert_weights.py
@@ -7,4 +7,4 @@ if __name__ == "__main__":
     for key, value in state_dict.items():
         print(f"Layer: {key}")
         print(f"Shape: {value.shape}")
-    np.savez("gpt2.npz", **{k: v.to(torch.float16).numpy() for k, v in state_dict.items()})
+    np.savez("gpt2.npz", **{k: v.to(torch.float32).numpy() for k, v in state_dict.items()})
